pyproject.toml
[tool.poetry]
name = "algoassessment"
version = "0.1.0"
description = "A server that searches for matches of a string given a configuration file"
authors = ["Kolawole <pro.kaiserlautern@gmail.com>"]
readme = "README.md"
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
fpdf = "^1.7.2"
pyahocorasick = "^2.1.0"
pandas-stubs = "^2.2.3.241126"

[tool.poetry.group.dev.dependencies]
# Linting and formatting
black = "^23.10.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
# Static type checking
mypy = "^1.6.1"
# Testing
pytest = "^7.4.2"
pytest-cov = "^4.1.0" # For coverage reports
# For docstrings
pydocstyle = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
include = '\.pyi?$'
exclude = '''
(
  ^(
    blib/
    |build/
    |buck-out/
    |dist/
    |eggs/
    |\.eggs/
    |lib/
    |lib64/
    |node_modules/
    |\.nox/
    |\.tox/
    |venv/
    |virtualenv/
  )/
)
'''
